ch4: uninline progress engine
Experiments has shown no performance impact after unlinine the entire
progress engine. This change significantly reduces the build size and
time.
