+++ src/mpid/ch4/netmod/ofi/func_table.c
@@ -20,7 +20,7 @@ MPL_SUPPRESS_OSX_HAS_NO_SYMBOLS_WARNING;
 MPIDI_NM_funcs_t MPIDI_NM_ofi_funcs = {
     .mpi_init = MPIDI_OFI_mpi_init_hook,
     .mpi_finalize = MPIDI_OFI_mpi_finalize_hook,
-    .progress = MPIDI_NM_progress,
+    .progress = MPIDI_OFI_progress,
     .mpi_comm_connect = MPIDI_OFI_mpi_comm_connect,
     .mpi_comm_disconnect = MPIDI_OFI_mpi_comm_disconnect,
     .mpi_open_port = MPIDI_OFI_mpi_open_port,
