+++ src/mpid/ch4/netmod/ofi/netmod_inline.h
@@ -14,7 +14,6 @@
 #include "ofi_am.h"
 #include "ofi_events.h"
 #include "ofi_proc.h"
-#include "ofi_progress.h"
 #include "ofi_unimpl.h"
 #include "ofi_coll.h"
 #include "ofi_probe.h"
