+++ src/mpid/ch4/netmod/stubnm/Makefile.mk
@@ -13,6 +13,7 @@ mpi_core_sources   += src/mpid/ch4/netmod/stubnm/func_table.c\
                       src/mpid/ch4/netmod/stubnm/stubnm_op.c \
                       src/mpid/ch4/netmod/stubnm/stubnm_spawn.c \
                       src/mpid/ch4/netmod/stubnm/stubnm_win.c \
+                      src/mpid/ch4/netmod/stubnm/stubnm_progress.c \
                       src/mpid/ch4/netmod/stubnm/globals.c
 # errnames_txt_files += src/mpid/ch4/netmod/stub/errnames.txt
