+++ src/mpid/ch4/netmod/stubnm/netmod_inline.h
@@ -11,7 +11,6 @@
 #ifndef NETMOD_INLINE_H_INCLUDED
 #define NETMOD_INLINE_H_INCLUDED
 #include "stubnm_probe.h"
-#include "stubnm_progress.h"
 #include "stubnm_recv.h"
 #include "stubnm_request.h"
 #include "stubnm_send.h"
