+++ src/mpid/ch4/shm/posix/shm_inline.h
@@ -11,7 +11,6 @@
 #define POSIX_INLINE_H_INCLUDED
 
 #include "posix_probe.h"
-#include "posix_progress.h"
 #include "posix_recv.h"
 #include "posix_request.h"
 #include "posix_send.h"
