+++ src/mpid/ch4/shm/src/func_table.c
@@ -21,7 +21,7 @@ MPIDI_SHM_funcs_t MPIDI_SHM_src_funcs = {
     .mpi_init = MPIDI_SHMI_mpi_init_hook,
     .mpi_finalize = MPIDI_SHMI_mpi_finalize_hook,
     .get_vci_attr = MPIDI_SHMI_get_vci_attr,
-    .progress = MPIDI_SHM_progress,
+    .progress = MPIDI_SHMI_progress,
     .mpi_comm_connect = MPIDI_SHMI_mpi_comm_connect,
     .mpi_comm_disconnect = MPIDI_SHMI_mpi_comm_disconnect,
     .mpi_open_port = MPIDI_SHMI_mpi_open_port,
