+++ src/mpid/ch4/src/Makefile.mk
@@ -13,7 +13,6 @@
 AM_CPPFLAGS += -I$(top_srcdir)/src/mpid/ch4/src
 
 noinst_HEADERS += src/mpid/ch4/src/ch4_comm.h     \
-                  src/mpid/ch4/src/ch4_progress.h \
                   src/mpid/ch4/src/ch4_request.h  \
                   src/mpid/ch4/src/ch4_send.h     \
                   src/mpid/ch4/src/ch4_types.h    \
@@ -43,6 +42,7 @@ mpi_core_sources += src/mpid/ch4/src/ch4_globals.c        \
                     src/mpid/ch4/src/ch4_init.c           \
                     src/mpid/ch4/src/ch4_comm.c           \
                     src/mpid/ch4/src/ch4_spawn.c          \
+                    src/mpid/ch4/src/ch4_progress.c       \
                     src/mpid/ch4/src/ch4_win.c            \
                     src/mpid/ch4/src/ch4i_comm.c          \
                     src/mpid/ch4/src/ch4r_buf.c           \
